// shipeasy/src/app/models/tds-slabs.ts
export interface TDS {
    tdsId: string;
    accountHead: string;
    createdBy: string;
    createdOn: string;
    educationPer: string;
    effectiveFrom: string;
    effectiveTo: string;
    entityType: string;
    module: string;
    orgId: string;
    remarks: string;
    sectionCode: string;
    status: boolean;
    surchargePer: string;
    tdsNature: string;
    tdsPer: string;
    tenantId: string;
    thresholdLimit: string;
    updatedBy: string;
    updatedOn: string;
    type: string;
    updatedByUID: string;
  }
  